mm: page_alloc: skip over regions of invalid pfns where possible
When using a sparse memory model memmap_init_zone() when invoked with the
MEMMAP_EARLY context will skip over pages which aren't valid - ie.  which
aren't in a populated region of the sparse memory map.  However if the
memory map is extremely sparse then it can spend a long time linearly
checking each PFN in a large non-populated region of the memory map &
skipping it in turn.

When CONFIG_HAVE_MEMBLOCK_NODE_MAP is enabled, we have sufficient
information to quickly discover the next valid PFN given an invalid one by
searching through the list of memory regions & skipping forwards to the
first PFN covered by the memory region to the right of the non-populated
region.  Implement this in order to speed up memmap_init_zone() for
systems with extremely sparse memory maps.

James said "I have tested this patch on a virtual model of a Samurai
CPU with a sparse memory map.  The kernel boot time drops from 109 to
62 seconds. " 

Link: http://lkml.kernel.org/r/20161125185518.29885-1-paul.burton@imgtec.com
Signed-off-by: Paul Burton <paul.burton@imgtec.com>
Tested-by: James Hartley <james.hartley@imgtec.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>