mm: fix register_new_memory() zone type detection
In preparation for sub-section memory hotplug support, remove a dependency
on ->section_mem_map being populated.  In SPARSEMEM_VMEMMAP=y
configurations pfn_to_page() does not use ->section_mem_map.  The
sub-section hotplug support relies on this fact and skips initializing it.
Without ->section_mem_map populated, or aligned to section boundary,
conversions of mem_section instances to zones is not possible.

So, this removes a false dependency on a structure field that will only be
valid in the SPARSEMEM_VMEMMAP=n case, and only used for pfn_to_page()
(and similar) operations.

Link: http://lkml.kernel.org/r/148486363923.19694.16803321290568634521.stgit@dwillia2-desk3.amr.corp.intel.com
Signed-off-by: Dan Williams <dan.j.williams@intel.com>
Cc: Michal Hocko <mhocko@suse.com>
Cc: Vlastimil Babka <vbabka@suse.cz>
Cc: Johannes Weiner <hannes@cmpxchg.org>
Cc: Logan Gunthorpe <logang@deltatee.com>
Cc: Mel Gorman <mgorman@techsingularity.net>
Cc: Stephen Bates <stephen.bates@microsemi.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>