userfaultfd-hugetlbfs-add-uffdio_copy-support-for-shared-mappings-update
Thanks Andrea, I incorporated your suggestions into a new version of the
patch.  While changing (dst_vma->vm_flags & VM_SHARED) to integers, I
noticed an issue in the error path of __mcopy_atomic_hugetlb().

>                */
> -             ClearPagePrivate(page);
> +             if (dst_vma->vm_flags & VM_SHARED)
> +                     SetPagePrivate(page);
> +             else
> +                     ClearPagePrivate(page);
>               put_page(page);

We can not use dst_vma here as it may be different than the vma for which
the page was originally allocated, or even NULL.  Remember, we may drop
mmap_sem and look up dst_vma again.  Therefore, we need to save the value
of (dst_vma->vm_flags & VM_SHARED) for the vma which was used when the
page was allocated.  This change as well as your suggestions are in the
patch below:

Link: http://lkml.kernel.org/r/c9c8cafe-baa7-05b4-34ea-1dfa5523a85f@oracle.com
Signed-off-by: Mike Kravetz <mike.kravetz@oracle.com>
Reviewed-by: Andrea Arcangeli <aarcange@redhat.com>
Cc: Mike Rapoport <rppt@linux.vnet.ibm.com>
Cc: Hillf Danton <hillf.zj@alibaba-inc.com>
Cc: Mike Kravetz <mike.kravetz@oracle.com>
Cc: Pavel Emelyanov <xemul@parallels.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>