kernel/ksysfs.c: add __ro_after_init to bin_attribute structure
The object notes_attr of type bin_attribute is not modified after getting
initailized by ksysfs_init.  Apart from initialization in ksysfs_init it
is also passed as an argument to the function sysfs_create_bin_file but
this argument is of type const.  Therefore, add __ro_after_init to its
declaration.

Link: http://lkml.kernel.org/r/1486839969-16891-1-git-send-email-bhumirks@gmail.com
Signed-off-by: Bhumika Goyal <bhumirks@gmail.com>
Acked-by: Kees Cook <keescook@chromium.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>