mm/put_page: move ref decrement to put_zone_device_page()
This does not affect non ZONE_DEVICE page.  In order to allow ZONE_DEVICE
page to be tracked we need to detect when refcount of a ZONE_DEVICE page
reach 1 (not 0 as non ZONE_DEVICE page).

This patch just move put_page_testzero() from put_page() to
put_zone_device_page() and only for ZONE_DEVICE.  It does not add any
overhead compare to existing code.

Link: http://lkml.kernel.org/r/1489680335-6594-3-git-send-email-jglisse@redhat.com
Signed-off-by: Jérôme Glisse <jglisse@redhat.com>
Cc: Dan Williams <dan.j.williams@intel.com>
Cc: Ross Zwisler <ross.zwisler@linux.intel.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>